Middleware: Simplify
