+++ pkg/registry/registry.go
@@ -25,7 +25,7 @@ func RegisterServiceWithPriority(instance Service, priority Priority) {
 }
 
 func RegisterService(instance Service) {
-	services = append(services, &Descriptor{
+	Register(&Descriptor{
 		Name:         reflect.TypeOf(instance).Elem().Name(),
 		Instance:     instance,
 		InitPriority: Medium,
@@ -33,6 +33,17 @@ func RegisterService(instance Service) {
 }
 
 func Register(descriptor *Descriptor) {
+	if descriptor == nil {
+		return
+	}
+	// Overwrite any existing equivalent service
+	for i, svc := range services {
+		if svc.Name == descriptor.Name {
+			services[i] = descriptor
+			return
+		}
+	}
+
 	services = append(services, descriptor)
 }
